requirements.txt
Django==5.1.5
PyPDF2==3.0.1
djangorestframework==3.15.2
mysql==0.0.3
pillow==11.1.0
python-docx==1.1.2